Step 3: extract cal_effective_D/newq/newd_* into vnl_pw_deeq.cpp
- Move cal_effective_D(), newq(), newd_so(), newd_nc() to vnl_pw_deeq.cpp
- Replace new[]/delete[] qnorm with std::vector in newq()
- Add doxygen comments for all functions
- Update CMakeLists.txt and Makefile.Objects
- Fix blas_connector.h include path (source_base/module_external/)

Verification:
  cd build_max_para_test && make -j 30
  => build passed
