+++ source/Makefile.Objects
@@ -723,6 +723,7 @@ OBJS_SRCPW=h_ewald_pw.o\
     dnrm2.o\
     vl_pw.o\
     vnl_pw.o\
+    vnl_pw_deeq.o\
     vnl_pw_getvnl.o\
     vnl_pw_grad.o\
     vnl_pw_qrad.o\
