+++ source/source_pw/module_pwdft/CMakeLists.txt
@@ -43,6 +43,7 @@ list(APPEND objects
     stress_pw.cpp
     vl_pw.cpp
     vnl_pw.cpp
+    vnl_pw_deeq.cpp
     vnl_pw_getvnl.cpp
     vnl_pw_grad.cpp
     vnl_pw_qrad.cpp
